refactor: create users-controller
