+++ src/main.js
@@ -1,115 +1,62 @@
 import 'dotenv/config';
 import http from 'http';
-import {v4 as uuid} from 'uuid';
 
 import {users} from './data.js';
+import * as UsersController from './usersController.js';
 
 const server = http.createServer((req, res) => {
   const urlPaths = req.url.split('/');
   urlPaths.splice(0, 1);
     
   if(urlPaths[0] === 'api' && urlPaths[1] === 'users' && req.method === 'GET'){
-    try {
-       // get user by id
+   try {
       if(urlPaths.length === 3){
-        const id = urlPaths[2];
-        const user = users.find((u) => u.id === id);
-      if(user){
-        res.end(JSON.stringify(user));
-      } else {
-        res.statusCode = 404;
-        res.end('404 - Not Found');
-      }
-      // get all users
+        UsersController.getUserById(urlPaths[2], res);
       } else if(urlPaths.length === 2){
-        res.end(JSON.stringify(users));
+        UsersController.getAllUsers(res);
       } else {
-        res.statusCode = 404;
-        res.end('404 - Not Found');
+        UsersController.userNotFound(res);
       }
    } catch(err){
-      res.statusCode = 500;
-      res.end('500 - Internal Server Error');
+      UsersController.internalServerError(res);
    }
-   
-  } else if(urlPaths[0] === 'api' && urlPaths[1] === 'users' && req.method === 'POST'){
     
+  } else if(urlPaths[0] === 'api' && urlPaths[1] === 'users' && req.method === 'POST'){
     req.on('data', (data) => {
       try{
-         const userData = JSON.parse(data.toString());
-         const newUser = {
-          id: uuid(),
-          ...userData
-         };
-         users.push(newUser);
-         res.statusCode = 201;
-         res.end(JSON.stringify(newUser));
+         UsersController.createUser(data, res);
       } catch(err){
-         res.statusCode = 500;
-         res.end('500 - Internal Server Error');
+         UsersController.internalServerError(res);
       } 
     });
     
     req.on('error', (error) => {
-      res.statusCode = 500;
-      res.end('500 - Internal Server Error');
+      UsersController.internalServerError(res);
     });
     
   } else if(urlPaths[0] === 'api' && urlPaths[1] === 'users' && urlPaths.length === 3 && req.method === 'PUT'){
-    
       req.on('data', (data) => {
         try{
-          const id = urlPaths[2];
-          const newUser = JSON.parse(data.toString());
-          const index = users.findIndex((user) => user.id === id);
-         
-          if(index !== -1){
-            const currentUser = users[index];
-            const updatedUser = {
-              ...currentUser,
-              ...newUser
-            };
-            users[index] = updatedUser;
-            res.end(JSON.stringify(updatedUser));   
-          } else {
-            res.statusCode = 404;
-            res.end('404 - Not Found');
-          }
+          UsersController.updateUser(urlPaths[2], data, res);
         } catch(err){
-          res.statusCode = 500;
-          res.end('500 - Internal Server Error');
+          UsersController.internalServerError(res);
         }
       });
       
      req.on('error', (error) => {
-      res.statusCode = 500;
-      res.end('500 - Internal Server Error');
+        UsersController.internalServerError(res);
      });
 
   } else if(urlPaths[0] === 'api' && urlPaths[1] === 'users' && urlPaths.length === 3  && req.method === 'DELETE'){
-    
     try {
-      const id = urlPaths[2];
-      const index = users.findIndex((user) => user.id === id);
-    
-      if(index !== -1){
-        users.splice(index, 1);
-        res.statusCode = 204;
-        res.end();
-      } else {
-        res.statusCode = 404;
-        res.end('404 - Not Found');
-      }
+      UsersController.deleteUser(urlPaths[2], res);
     } catch(err){
-        res.statusCode = 500;
-        res.end('500 - Internal Server Error');
+      UsersController.internalServerError(res);
     }
   
   } else {
-    res.statusCode = 404;
-    res.end('404 - Not Found');
+    UsersController.userNotFound(res);
   }
-  
 });
 
 server.listen(process.env.PORT, () => {
